requirements.txt
jsonlines ~= 4.0.0
numpy >= 1.26.4, < 2.0.0
openai ~= 1.57.1
pandas ~= 2.0.3
pdoc ~= 15.0.0
pillow ~= 10.3.0
pyheaven >= 0.1.6.4
send2trash ~= 1.8.0
Wand ~= 0.6.13
